rename network.py to networks.py
