+++ starsim/networks.py
